requirements.txt
appium-python-client==2.9.0
selenium==4.12.0
pytest==7.4.0
requests
appium-python-client
selenium

